Refactor Converter services file
