--- CONVERT_SERVICES/src/controller/backup/rest_api_dayler/.gitignore
@@ -1,2 +1,0 @@
-files/*
-**/__pycache__/
