--- CONVERT_SERVICES/src/controller/backup/rest_api_dayler/requirements.txt
@@ -1,2 +1,0 @@
-Flask==2.0.3
-Flask-RESTful==0.3.9
