--- CONVERT_SERVICES/src/controller/backup/rest_api_dayler/resources/file_uploader.py
@@ -1,31 +1,0 @@
-#
-# @main.py Copyright (c) 2022 Jalasoft.
-# 2643 Av Melchor Perez de Olguin, Colquiri Sud, Cochabamba, Bolivia.
-# Edificio Union № 1376 Av. General Inofuentes esquina Calle 20, La Paz, Bolivia.
-# All rights reserved.
-#
-# This software is the confidential and proprietary information of
-# Jalasoft, ("Confidential Information"). You shall not
-# disclose such Confidential Information and shall use it only in
-# accordance with the terms of the license agreement you entered into
-# with Jalasoft.
-#
-
-import werkzeug
-from flask_restful import Resource, reqparse
-
-
-class FileUploader(Resource):
-    def post(self):
-        parse = reqparse.RequestParser()
-        parse.add_argument('file', type=werkzeug.datastructures.FileStorage, location='files')
-        parse.add_argument('name', type=str)
-        args = parse.parse_args()
-        image_file = args['file']
-        name = args['name']
-        filename = f"{name}.jpg"
-        image_file.save(f"files/{filename}")
-        return {"file_id": filename}
-
-    def get(self):
-        pass
