--- CONVERT_SERVICES/src/pdf_to_word/obtName.py
@@ -1,6 +1,0 @@
-class ObtName():
-    def __init__(self, name):
-        self.name = name
-
-    def get_name(self):
-        return self.name
