+++ MACHINE_LEARNING_SERVICE/src/controller/utils/zipfile/zip.py
@@ -1,5 +1,5 @@
 #
-# @main.py Copyright (c) 2022 Jalasoft.
+# @converter_pdf_to_word.py Copyright (c) 2022 Jalasoft.
 # 2643 Av Melchor Perez de Olguin, Colquiri Sud, Cochabamba, Bolivia.
 # Edificio Union № 1376 Av. General Inofuentes esquina Calle 20, La Paz, Bolivia.
 # All rights reserved.
